Move guard functions to top level
