+++ src/compiler/typescript_to_rust/azle_generate/src/guard_function/rust.rs
